+++ src/Components/Cards/Camera/Camera.js
@@ -5,7 +5,7 @@ import { withStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import ButtonBase from '@material-ui/core/ButtonBase';
 import Card from '@material-ui/core/Card';
-import { getCardElevation, getSquareCards } from '../../Utils/config';
+import { getCardElevation, getSquareCards } from '../../Common/config';
 import Dialog from './Dialog';
 
 const styles = theme => ({
