+++ src/Components/Common/clone.js
